+++ SKO_Network/SKO_Network.h
@@ -27,7 +27,7 @@ class SKO_Network
  
 public:
 
-	SKO_Network(SKO_Repository * repository, int port, unsigned long int saveRateSeconds);
+	SKO_Network(SKO_Repository * repository, int port, unsigned long long int saveRateSeconds);
 	std::string startup();
 	void cleanup();
 
@@ -149,7 +149,7 @@ public:
 	std::mutex saveMutex;
 
 	// How often to save all valid players.
-	unsigned long int saveRateSeconds;
+	unsigned long long int saveRateSeconds;
 	unsigned long long int lastSaveTime;
 	
 	// Do not spam the database status table if there have been 0 players online for some time.
